refactor: move Header and Layout components to separate folders
